requirements.txt
FastAPI>=0.100.0
uvicorn>=0.24.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
fastapi-cors>=0.0.6
bcrypt>=4.0.0
cryptography>=41.0.0